Change interface to copy all blocks
Digging into the project files, I realized that copying blocks by name
is... tricky.  The obvious initial case we want is to just copy all of
the blocks in the project file to the new project file.  So I reworked
the standard interface to instead copy all of the blocks, and not ask
for a block name to copy.